insert call for fundCalculation
the calculation get's called by for-next loop
as often as the LAUFZEIT was set in code,
and given the interest rates set in code.

The new fund get's printed after each calculation
and on print calculated back from Cent to Euro
only in the print, keeping the variable value
always in Cent.